requirements.txt
astroid==2.2.5
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.2
wrapt==1.11.2
gunicorn==19.9.0

